Fixes dashboards which broke if they reassigned $start
If a dashboard had a query like
```
$s = $start
q(..something..., "$s")
```
This broke as $start was subbed in with quotes, leading to double
quotes and an error response. The original changes that this reverts
were to designed remove the requirement for quotes around $start,
but since this is the behaviour in many places, it probably makes
sense to leave it as is and require the quotes.